Wire up autocomplete component, add actions to handle location change
This action will search for city names based on user input. I also added debounce and throttle utility methods as part of this commit. Since the search request is made on user input, it needed to be debounced to prevent making excessive requests. It is currently debounced at a rate of 500ms.

We could bring in lodash to accomplish this, but I don't know if we'll need much more than these