Create selectedCity vuex property, set when user selects a city
We needed a way to display the selected city in the card header. I figured the easiest way to accomplish this would be to set the selected city at the same time as the request to fetch the selected city's weather.

In addition to creating this property, it is also used in card header.