Refactored Vuex store, in particular actions
It is advisable for actions to return promises. It makes chaining several actions together easier, since the actions themselves can be dispatched in one function that awaits the promises each action returns.

This also adds error handling to our requests. At this point it is only logging to the console.error, but we can develop our error handling/messaging later.