Create utilities file for parsing; extract some Vuex logic there
We want to limit the amount of parsing logic in our Vuex store. That way this logic can be more easily isolated and tested, as well as reused elsewhere throughout the app; in addition, the Vuex store can have the sole responsibility of managing state, not parsing XHR data.

This commit also includes a few small fixes and optimizations.